+++ nodebuilder/daser/module.go
@@ -8,7 +8,7 @@ import (
 	"github.com/celestiaorg/celestia-node/das"
 	"github.com/celestiaorg/celestia-node/fraud"
 	"github.com/celestiaorg/celestia-node/libs/fxutil"
-	fraudbuilder "github.com/celestiaorg/celestia-node/nodebuilder/fraud"
+	fraudServ "github.com/celestiaorg/celestia-node/nodebuilder/fraud"
 	"github.com/celestiaorg/celestia-node/nodebuilder/node"
 )
 
@@ -19,9 +19,9 @@ func Module(tp node.Type) fx.Option {
 			"daser",
 			fx.Provide(fx.Annotate(
 				NewDASer,
-				fx.OnStart(func(ctx context.Context, lc fx.Lifecycle, fservice fraud.Service, das *das.DASer) error {
+				fx.OnStart(func(ctx context.Context, lc fx.Lifecycle, fservice fraudServ.Service, das *das.DASer) error {
 					lifecycleCtx := fxutil.WithLifecycle(ctx, lc)
-					return fraudbuilder.Lifecycle(ctx, lifecycleCtx, fraud.BadEncoding, fservice,
+					return fraudServ.Lifecycle(ctx, lifecycleCtx, fraud.BadEncoding, fservice,
 						das.Start, das.Stop)
 				}),
 				fx.OnStop(func(ctx context.Context, das *das.DASer) error {
